# public/api-docs.yaml
openapi: 3.0.3
info:
  title: Assessment API - Video Platform
  description: A comprehensive video sharing platform API with authentication, video management, social features, and content discovery
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000/api
    description: Development server
security:
  - JWT: []
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        errors:
          type: array
          items:
            type: string
        pagination:
          type: object
paths:
...
  /login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - username
                - password
              properties:
                email:
                  type: string
                  format: email
                username:
                  type: string
                firstName:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Registration successful
  /refresh:
    post:
      tags:
        - Authentication
      summary: Refresh JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
  /logout:
    post:
      tags:
        - Authentication
      summary: User logout
      security:
        - JWT: []
      responses:
        '200':
          description: Logged out successfully
  /me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      security:
        - JWT: []
      responses:
        '200':
          description: User profile retrieved

  # Video Management Endpoints
...
  /feed/vertical:
    get:
      tags:
        - Feed
      summary: Get vertical feed
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Vertical feed retrieved
  /feed/home:
    get:
      tags:
        - Feed
      summary: Get home feed
      security:
        - JWT: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Home feed retrieved
  /feed/trending:
    get:
      tags:
        - Feed
      summary: Get trending feed
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Trending feed retrieved
  /feed/chronological:
    get:
      tags:
        - Feed
      summary: Get chronological feed
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Chronological feed retrieved
  /feed/creator/{userId}:
    get:
      tags:
        - Feed
      summary: Get creator channel
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Creator channel retrieved
  /feed/discover:
    get:
      tags:
        - Feed
      summary: Get discover feed
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
            default: all
      responses:
        '200':
          description: Discover feed retrieved

  # Search Endpoints
...